third_party/blink/web_tests/platform/generic/fast/css/counters/counter-increment-not-reflected-on-stylechange-expected.txt

PASS internals.counterValue(document.getElementById('inner1')) is "1"
PASS internals.counterValue(document.getElementById('inner2')) is "2"
PASS successfullyParsed is true

TEST COMPLETE
